refactor: statically render canto gallery pages
