+++ app/[locale]/gallery/collections/[gallery]/page.tsx
@@ -20,6 +20,4 @@ const Gallery: FunctionComponent<WithSearchParams<GalleryProps>> = async ({
   return <GalleryPage {...{ locale, gallery, searchParams }} />;
 };
 
-export const dynamic = "force-dynamic";
-
 export default Gallery;
